+++ lib/clixon/clixon_xml.h
@@ -115,7 +115,7 @@ typedef struct clixon_xml_vec clixon_xvec; /* struct defined in clicon_xvec.c */
 #define XML_FLAG_DEL    0x04  /* Node is deleted (commits) or parent deleted rec */
 #define XML_FLAG_CHANGE 0x08  /* Node is changed (commits) or child changed rec */
 #define XML_FLAG_NONE   0x10  /* Node is added as NONE */
-#define XML_FLAG_DEFAULT 0x20 /* Added as default value @see xml_default*/
+#define XML_FLAG_DEFAULT 0x20 /* Added as default value @see xml_default */
 
 /*
  * Prototypes